# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateOrderRequest:
    properties:
      product_id:
        type: integer
    type: object
  handler.CreateOrderResponse:
    properties:
      message:
        type: string
    type: object
  handler.CreateProductRequest:
    properties:
      name:
        example: Product Name
        maxLength: 100
        minLength: 2
        type: string
      price_cents:
        example: 12000
        type: integer
    required:
    - name
    type: object
  handler.CreateProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
  handler.DeleteOrderResponse:
    properties:
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.GetOrderResponse:
    properties:
      item:
        $ref: '#/definitions/handler.OrderItem'
    type: object
  handler.GetOrdersResponse:
    properties:
      hasNext:
        type: boolean
      items:
        items:
          $ref: '#/definitions/handler.OrderItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handler.GetProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
  handler.GetProductsResponse:
    properties:
      hasNext:
        type: boolean
      items:
        items:
          $ref: '#/definitions/handler.ProductItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handler.LoginUserRequest:
    properties:
      password:
        example: password
        maxLength: 100
        minLength: 2
        type: string
      user_name:
        example: admin
        maxLength: 100
        minLength: 2
        type: string
    required:
    - password
    - user_name
    type: object
  handler.LoginUserResponse:
    properties:
      access_token:
        type: string
      token_type:
        type: string
    type: object
  handler.OrderItem:
    properties:
      customer_name:
        type: string
      id:
        type: integer
      price_cents:
        type: integer
      product_name:
        type: string
    type: object
  handler.ProductItem:
    properties:
      id:
        type: integer
      name:
        type: string
      price_cents:
        type: integer
    type: object
  handler.RegisterUserRequest:
    properties:
      password:
        example: my-password
        maxLength: 100
        minLength: 2
        type: string
      user_name:
        example: My Name
        maxLength: 100
        minLength: 2
        type: string
    required:
    - password
    - user_name
    type: object
  handler.RegisterUserResponse:
    properties:
      status:
        type: string
    type: object
  handler.UpdateOrderRequest:
    properties:
      price_cents:
        type: integer
    type: object
  handler.UpdateOrderResponse:
    properties:
      item:
        $ref: '#/definitions/handler.OrderItem'
    type: object
  handler.UpdateProductRequest:
    properties:
      name:
        example: Product Name
        maxLength: 100
        minLength: 2
        type: string
      price_cents:
        example: 12000
        type: integer
    type: object
  handler.UpdateProductResponse:
    properties:
      item:
        $ref: '#/definitions/handler.ProductItem'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample car ordering server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend Example API
  version: "1.0"
paths:
  /login-user:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token.
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: User Login
      tags:
      - auth
  /orders:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of orders. Requires Admin role.
      parameters:
      - description: field to sort by (id, price_cents, created_at)
        in: query
        name: orderBy
        type: string
      - description: sort direction (asc, desc)
        in: query
        name: sortIn
        type: string
      - description: comma-separated product IDs to filter
        in: query
        name: product_ids
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Places an order for a specific product. Requires Customer role.
      parameters:
      - description: Order Payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an order by ID (Admin only).
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Returns a single order. Admins can view any order; customers can
        view their own orders.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an order by ID
      tags:
      - orders
    patch:
      consumes:
      - application/json
      description: Updates an order's fields (Admin only).
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an order
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of products with optional filtering and
        sorting.
      parameters:
      - description: field to sort by (name, price_cents)
        in: query
        name: orderBy
        type: string
      - description: sort direction (asc, desc)
        in: query
        name: sortIn
        type: string
      - description: filter by product name (partial match)
        in: query
        name: name
        type: string
      - description: minimum price in cents
        in: query
        name: minPrice
        type: integer
      - description: maximum price in cents
        in: query
        name: maxPrice
        type: integer
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product (Admin only).
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a product
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Returns a single product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a product by ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Updates product fields (Admin only).
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
  /register-admin:
    post:
      consumes:
      - application/json
      description: Creates a new user with the Admin role.
      parameters:
      - description: Admin Registration Payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register an Admin
      tags:
      - auth
  /register-customer:
    post:
      consumes:
      - application/json
      description: Creates a new user with the Customer role.
      parameters:
      - description: Customer Registration Payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a Customer
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"